+++ ovos_utils/skills/templates/auto_translatable.py
